Fixed: allow normal class extension for plugin classes
Before this fix, having a plugin class extend its own plugin
class (from a different file) would cause the compiler to fail to
resolve the method calling order.

This is now resolved, by excluding base classes from the actual plugin
class, which is created on the fly by yaz, when it is already
implicitly included as a parent class.

Fixes #4